requirements.txt
scrapy
pandas
ipython
scrapy-rotating-proxies